% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.twowayfeweights}
\alias{print.twowayfeweights}
\title{A print method for twowayfeweights objects}
\usage{
\method{print}{twowayfeweights}(x, ...)
}
\arguments{
\item{x}{A twowayfeweights object.}

\item{...}{Currently ignored.}
}
\description{
Printed display of twowayfeweights objects.
}
\examples{
# The following example is based on data from F. Vella and M. Verbeek (1998), 
# "Whose Wages Do Unions Raise? A Dynamic Model of Unionism and Wage Rate 
# Determination for Young Men". 
# Results and further details about the estimation below can be found in 
# Section V.C of de Chaisemartin & D'Haultfoeuille (2020a).
# Run the following lines to download the dataset in your local working 
# directory and load it to your R environment:

# install.packages("haven")
repo = "chaisemartinPackages/twowayfeweights/main"
file = "wagepan_twfeweights.dta"
url = paste("https://raw.githubusercontent.com", repo, file, sep = "/")
wagepan =  haven::read_dta(url)

# Now let's load this package and run some estimations.
library(TwoWayFEWeights)

# The default `type = "feTR"` estimation strategy uses a fixed-effects
# strategy under the assumption that parallel trends holds. 
twowayfeweights(
  wagepan,
  "lwage", "nr", "year", "union",
  type = "feTR", # default
  summary_measures = TRUE,
  test_random_weights = "educ"
)

# The next line performs the same exercise using first differences of outcome
# and treatment:
twowayfeweights(
  wagepan,
  "diff_lwage", "nr", "year", "diff_union", 
  type = "fdTR",
  D0 = "union",
  summary_measures = TRUE,
  test_random_weights = "educ"
)

# Please note that the number of negative weights could be different from Section 
# V.C. of de Chaisemartin and D'Haultfoeuille (2020a) due to rounding errors that 
# affected older versions of the commands.

}
